requirements.txt
websocket-client
coloredlogs
aioconsole
pyyaml
